# Conflicts:
#	Four Peaks/AlphaTemp/collision.cpp

+++ Four Peaks/AlphaTemp/collision.hpp
@@ -4,7 +4,7 @@
 #include <string>
 extern int (*g_currentMap)[32];
 extern int g_currentSignID;
-extern int g_currentY;
+extern float g_currentY;
 extern std::string g_currentScene;
 
 void CollisionUpdate(Player &player,f32 dt);
